requirements.txt
# PDF Processing
PyPDF2
pdfplumber

# Document Processing
python-docx
openpyxl
pandas

# Text Search
rank_bm25

# Web Framework
fastapi
uvicorn[standard]
pydantic
python-multipart
aiofiles

# Utilities
python-dotenv
colorama

# Rate Limiting
slowapi

# Testing
pytest
httpx

# Database
psycopg2-binary
sqlalchemy
alembic
requests

# LLM
openai
tiktoken
